MB-29119: Replace revSeqno with a 48-bit counter
Prevent a value too large to be stored in couchstore
from being placed into Item/StoredValue and also the
_local document (via vbucket_state).

Reviewed-on: http://review.couchbase.org/92485
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>

+++ engines/ep/src/ep_bucket.cc
@@ -308,7 +308,7 @@ int EPBucket::flushVBucket(uint16_t vbid) {
                     vbstate.maxCas = std::max(vbstate.maxCas, item->getCas());
                     if (item->isDeleted()) {
                         vbstate.maxDeletedSeqno =
-                                std::max(vbstate.maxDeletedSeqno,
+                                std::max(uint64_t(vbstate.maxDeletedSeqno),
                                          item->getRevSeqno());
                     }
                     ++stats.flusher_todo;
